pyproject.toml
[tool.poetry]
name = "python-flask-example"
version = "0.1.0"
description = "A simple Flask app"
authors = ["Hexlet Team <info@hexlet.io>"]
license = "ISC"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^3.0.3"
gunicorn = "^23.0.0"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
